+++ tests/test_async.py
@@ -14,7 +14,7 @@ pytestmark = pytest.mark.anyio
 
 
 @pytest.mark.parametrize("attempts", [None, -1, 0, 1])
-@pytest.mark.parametrize("timeout", [None, -1, 0, 1, dt.timedelta(days=1)])
+@pytest.mark.parametrize("timeout", [None, -1, 1, dt.timedelta(days=1)])
 @pytest.mark.parametrize("duration", [1, dt.timedelta(days=1)])
 async def test_ok(attempts, timeout, duration):
     """